move kill_process_by_pid to tests/common/utilities.py
